+++ refena/lib/src/proxy_ref.dart
@@ -85,7 +85,7 @@ class ProxyRef implements Ref {
 
   @override
   void dispose<N extends BaseNotifier<T>, T>(BaseProvider<N, T> provider) {
-    _ref.dispose<N, T>(provider);
+    _ref.internalDispose<N, T>(provider, _debugOriginRef);
   }
 
   @override