+++ refena/test/unit/notifier/base_notifier_test.dart
@@ -166,6 +166,5 @@ class _DisposableCounter extends Notifier<int> {
   @override
   void dispose() {
     onDispose();
-    super.dispose();
   }
 }